refactor(core): centralize normalizeSince in src/core/time.ts as parseSinceToIso
Fixes #318
